V 1.2.0 Improved synchronization behavior for Commands
Modifed the injection point for new commands which should enable most
commands to be visible to an unmodded client and allow proper
syncronization with unmodded clients entering commands.

The behavior of the "Press any key to continue" block should behave as
it did with the Vanilla terminal. This should prevent odd buffering
behaviors on commands that have relatively short outputs and will be
most noticable when approaching a terminal that another mod use was
extensively using.